actions/git-path-filter/requirements.txt
# PyYAML is required for parsing the YAML-formatted pattern_filter input.
PyYAML>=6.0
